# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Dokümantasyonu",
    "version": "1.0.0",
    "description": "Express API dokümantasyonu"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Yeni kullanıcı kaydı",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kullanıcı başarıyla oluşturuldu"
          },
          "400": {
            "description": "Geçersiz input veya kullanıcı zaten mevcut"
          },
          "500": {
            "description": "Sunucu hatası"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Kullanıcı girişi",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı giriş",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Geçersiz kimlik bilgileri"
          },
          "500": {
            "description": "Sunucu hatası"
          }
        }
      }
    },
    "/api/todos": {
      "get": {
        "summary": "Kullanıcının tüm todo'larını getirir",
        "tags": [
          "Todos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Todo listesi başarıyla getirildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "completed": {
                        "type": "boolean"
                      },
                      "userId": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Kullanıcı girişi yapılmamış"
          },
          "500": {
            "description": "Sunucu hatası"
          }
        }
      },
      "post": {
        "summary": "Yeni bir todo oluşturur",
        "tags": [
          "Todos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo başarıyla oluşturuldu"
          },
          "400": {
            "description": "Geçersiz istek (title eksik)"
          },
          "401": {
            "description": "Kullanıcı girişi yapılmamış"
          }
        }
      }
    },
    "/api/todos/{id}": {
      "put": {
        "summary": "Bir todo'yu günceller",
        "tags": [
          "Todos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Güncellenecek todo'nun ID'si"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "completed": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo başarıyla güncellendi"
          },
          "401": {
            "description": "Kullanıcı girişi yapılmamış"
          },
          "404": {
            "description": "Todo bulunamadı"
          }
        }
      },
      "delete": {
        "summary": "Bir todo'yu siler",
        "tags": [
          "Todos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Silinecek todo'nun ID'si"
          }
        ],
        "responses": {
          "200": {
            "description": "Todo başarıyla silindi"
          },
          "401": {
            "description": "Kullanıcı girişi yapılmamış"
          },
          "404": {
            "description": "Todo bulunamadı"
          }
        }
      }
    },
    "/api/users/profile": {
      "get": {
        "summary": "Kullanıcı profilini getir",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Kullanıcı profili",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Yetkilendirme gerekli"
          },
          "404": {
            "description": "Kullanıcı bulunamadı"
          }
        }
      }
    },
    "/api/users/password": {
      "put": {
        "summary": "Kullanıcı şifresini güncelle",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Şifre başarıyla güncellendi"
          },
          "401": {
            "description": "Geçersiz mevcut şifre"
          },
          "404": {
            "description": "Kullanıcı bulunamadı"
          }
        }
      }
    },
    "/api/users/delete-account": {
      "delete": {
        "summary": "Kullanıcı hesabını sil",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hesap başarıyla silindi"
          },
          "401": {
            "description": "Geçersiz şifre"
          },
          "404": {
            "description": "Kullanıcı bulunamadı"
          }
        }
      }
    }
  },
  "tags": []
}